% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{accessors}
\alias{accessors}
\alias{get_series}
\alias{get_start}
\alias{get_state}
\alias{get_update}
\alias{get_wrapper}
\alias{is_exhausted}
\alias{is_limited}
\title{Accessor methods for the generator class}
\usage{
get_start(.generator)

get_state(.generator)

get_series(.generator)

get_update(.generator)

get_wrapper(.generator)

is_limited(.generator)

is_exhausted(.generator)
}
\arguments{
\item{.generator}{A function with the class \code{generator}}
}
\description{
\describe{
 \item{\code{get_start}}{Get the starting state of a generator}
 \item{\code{get_state}}{Get the current state of a generator}
 \item{\code{get_series}}{Get the current series value of a generator}
 \item{\code{get_update}}{Get the updating function of a generator}
 \item{\code{get_wrapper}}{Get the wrapper/ yield function of a generator}
 \item{\code{is_limited}}{Binary: show if the generator has a limiting
     function}
 \item{\code{is_exhausted}}{Binary: show if the generator is exhausted}
}
}
\details{
These functions make it possible to access various aspects of the generator
class.
}

