% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consume.R
\name{consume}
\alias{consume}
\title{Consume an generator, taking all iterations up to a limit}
\usage{
consume(.generator, .guess, .simplify)
}
\arguments{
\item{.generator}{A function with the class \code{generator}}

\item{.guess}{An approximate number of steps until the limit}

\item{.simplify}{If \code{FALSE} return a list, else return an array}
}
\description{
\code{consume} is a special version of \code{take}, with additional controls
to help the user. It produces the same output and warns the user (through
guess) if the number of iterations taken is insufficient to reach the limit.
}
\details{
It needs to be preceded by a \code{limit} to a generator.
The function checks whether a generator is limited, but other errors are
possible if the limit occurs earlier in the chain of wrappers.
}
\examples{
# A typical example
counter <- generator(0, function (state) state + 1)
counter \%>\% limit(~ state < 20) \%>\% consume

# Should throw a warning
\dontrun{
counter \%>\%
  reset \%>\%
  limit(~ state < 20) \%>\%
  consume(10)
}

}

