% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy.R
\name{copy}
\alias{copy}
\title{Copy a generator safely}
\usage{
copy(.generator)
}
\arguments{
\item{.generator}{A function with the class \code{generator}}
}
\description{
Assigning a new name to a generator does not break the link with the
underlying state. Calling the new generator will update the state in the
previous too. This is an unsafe copy. To break this link, and spawn a new
underlying state, use this function.
}
\examples{
# Unsafe copy
counter <- generator(0, ~ state + 1)
counter2 <- counter
counter2(); counter(); counter()
identical(counter, counter2)

# Safe copy
counter2 <- copy(counter)
counter2(); counter(); counter()
identical(counter, counter2)

}

