% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{discard}
\alias{discard}
\title{Filter a series by discarding results that fulfill a predicate}
\usage{
discard(.generator, .drop)
}
\arguments{
\item{.generator}{A function with the class \code{generator}}

\item{.drop}{A predicate function, returning \code{TRUE} or \code{FALSE}}
}
\description{
Filter a series by discarding results that fulfill a predicate
}
\examples{
# Generate even numbers
counter <- generator(0, ~ state + 1)
counter \%>\% discard(~ state \%\% 2 == 0) \%>\% take(5)

# Combine a wrap and a counter
squares <- function(state) state ^ 2
counter \%>\% wrap(squares) \%>\% discard(~ state \%\% 3 == 0) \%>\% take(5)

}

