% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\docType{package}
\name{generators}
\alias{generators}
\alias{generators-package}
\title{Generators: Iterable functions}
\description{
Generators are functions that contain a mutable internal state.
Calls of the function modify that state, making the function iterable:
repeated calls of the function return different values. When used correctly,
generators offer an efficient solution to many computational
problems. This package provides a grammar for working with these tools.
}
\details{
This package was inspired by Reginald Braithwaite's book on
Javascript
\href{https://leanpub.com/javascriptallongesix}{Javascript Allonge} and
Jeff Knupp's blog post on
\href{https://jeffknupp.com/blog/2013/04/07/improve-your-python-yield-and-generators-explained/}{using generators in Python}.
It also borrows heavily from
\href{https://doc.rust-lang.org/std/iter/trait.Iterator.html}{Rust's iterators methods},
which inspired
\href{https://doc.rust-lang.org/book/iterators.html}{the framework of iterator adaptors and consumers}.

Generators behave much like iterators, which are objects with a mutable
internal state. There are multiple implementations of iterators in R. See
below.
}
\seealso{
\href{https://cran.r-project.org/web/packages/iterators/index.html}{iterators}
and
\href{https://cran.r-project.org/web/packages/itertools/index.html}{itertools}
}

