% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{keep}
\alias{keep}
\title{Filter a series by keeping results that fulfill a predicate}
\usage{
keep(.generator, .keep)
}
\arguments{
\item{.generator}{A function with the class \code{generator}}

\item{.keep}{A predicate function, returning \code{TRUE} or \code{FALSE}}
}
\description{
Filter a series by keeping results that fulfill a predicate
}
\examples{
# Generate even numbers
counter <- generator(0, ~ state + 1)
counter \%>\%
  keep(~ state \%\% 3 == 0) \%>\% take(5)

# Combine a wrap and a counter
squares <- function(state) state ^ 2
counter \%>\%
  wrap(squares) \%>\%
  keep(~ state \%\% 3 == 0) \%>\%
  take(5)

}

