% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peek.R
\name{peek}
\alias{peek}
\title{Safely check for next value in iterator}
\usage{
peek(.generator)
}
\arguments{
\item{.generator}{A function with the class \code{generator}}
}
\description{
Safely check for next value in iterator
}
\examples{
counter <- generator(0, ~ state + 1)
peek(counter)
counter()

}

