% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebase.R
\name{rebase}
\alias{rebase}
\title{Alter a generators update function, spawning a new generator}
\usage{
rebase(.generator, .FUN, .wrapper = get_wrapper(.generator))
}
\arguments{
\item{.generator}{A function with the class \code{generator}}

\item{.FUN}{A function to wrap the current \code{update}}

\item{.wrapper}{A function to replace the current wrapper}
}
\description{
In general, it's not very good idea to alter the update of an generator.
If a different update is needed, you can always write a new generator, and
if one is needed in a chain of calls, \code{\link{wrap}}is usually a better
option.
}
\details{
That said, when used appropriately, rebasing is very useful. For example, it
powers several other tools in this package, like \code{\link{recycle}}.
}
\examples{
counter <- generator(0, ~ state + 1)
times_two <- function(x) x * 2
skipping <- rebase(counter, times_two)
take(skipping, 5)

}

