% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limit.R
\name{recycle}
\alias{recycle}
\title{Recycle a generator}
\usage{
recycle(.generator, .limit)
}
\arguments{
\item{.generator}{A function with the class \code{generator}}

\item{.limit}{A predicate function, returning \code{TRUE} or \code{FALSE}}
}
\description{
Recycling is limiting with one major change. Instead of throwing an error
when the generator is exhuasted, \code{recycle} resets the state of the
generator to its starting value.
}
\examples{
counter <- generator(0, function (state) state + 1)
recycled <- counter \%>\% recycle(~ state < 4)
recycled \%>\% take(7)

}

