% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset.R
\name{reset}
\alias{reset}
\title{Return a generator to its original state}
\usage{
reset(.generator)
}
\arguments{
\item{.generator}{A function with the class \code{generator}}
}
\description{
Return a generator to its original state
}
\examples{
counter <- generator(0, ~ state + 1)
take(counter, 5)
reset(counter)
counter()

}

