% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skip.R
\name{skip}
\alias{skip}
\title{Skip ahead in a generator}
\usage{
skip(.generator, .n)
}
\arguments{
\item{.generator}{A function with the class \code{generator}}

\item{.n}{The number to steps to move forward}
}
\description{
Skip ahead in a generator
}
\examples{
counter <- generator(0, ~ state + 1)
skip(counter, 5)
counter()

}

