% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/take.R
\name{take}
\alias{take}
\title{Extract n steps from a generator}
\usage{
take(.generator, .n, .simplify)
}
\arguments{
\item{.generator}{A function with the class \code{generator}}

\item{.n}{The number of steps to take}

\item{.simplify}{If \code{TRUE}, let the function try to simplify the output
instead of returning a list}
}
\description{
\code{.simplify} uses \code{unlist} and reduces lists to vectors in a
similar fashion.
}
\examples{
# Simplifying to a vector
counter <- generator(0, ~ state + 1)
take(counter, 5)

# Simplifying to a matrix
ran <- generator(c(0, 0), ~ rnorm(2))
take(ran, 5)

# Returning a list
take(ran, 5, FALSE)

}

