% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.R
\name{wrap}
\alias{wrap}
\title{Modify a generators wrapping function, spawning a new generator}
\usage{
wrap(.generator, .wrap)
}
\arguments{
\item{.generator}{A function with the class \code{generator}}

\item{.wrap}{A function that takes \code{state} as an argument}
}
\description{
Wrapping is the standard technique for spawning new generators. By default
all generators are wrapped in the \code{identity} function. When applying
\code{wrap}, we take the value currently returned by the generator and
use it as the argument in another function. To be safe, the function spawns
a new generator. Updating the second generator does not affect the first.
}
\details{
Wrapping is also the engine for a variety of other standard modifications
to generators, including filtering.
}
\examples{
# Create a counter than returns the square values of the natural numbers
counter <- generator(0, function (state) state + 1)
squares <- function(state) state ^ 2
squared <- counter \%>\% wrap(squares)
squared \%>\% take(5)

}

