% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\docType{package}
\name{generators}
\alias{generators}
\alias{generators-package}
\title{Generators: A grammar of iterators}
\description{
Iterators are a type of object that contains an internal state.
When used correctly, iterators offer an efficient solution to many computational
problems. This package provides a grammar for working with these tools.
}
\seealso{
\url[iterators]{https://cran.r-project.org/web/packages/iterators/index.html} and \url[itertools]{https://cran.r-project.org/web/packages/itertools/index.html}
}

